# feat: check for " " values and return a static instance