Add a check to `Doc`'s implicit operator for " " returning a static instance if found. Saves 0.5 MB in tests aka around 0.5%

### Benchmarks
#### Before
| Method                | Mean     | Error   | StdDev   | Gen0       | Gen1      | Gen2      | Allocated |
|---------------------- |---------:|--------:|---------:|-----------:|----------:|----------:|----------:|
| Default_CodeFormatter | 238.8 ms | 4.75 ms | 11.92 ms | 11000.0000 | 4000.0000 | 1000.0000 |  95.82 MB |


#### After
| Method                | Mean     | Error   | StdDev   | Gen0       | Gen1      | Gen2      | Allocated |
|---------------------- |---------:|--------:|---------:|-----------:|----------:|----------:|----------:|
| Default_CodeFormatter | 236.6 ms | 4.72 ms | 11.32 ms | 11000.0000 | 4000.0000 | 1000.0000 |  95.27 MB |