@TimothyMakkison 2025-02-19
See #1494